Create RCode.R
This code is meant to solve a classification problem. It selects numeric variables free of missing values from a dataset with a minimum correlation of 0.1 with the target variable; fits three kinds of models on the training data and applies them to preliminary test data; and given that the XGBoost performs best, that model is used to create a final, validation-set prediction.